+++ pyogrio/tests/test_core.py
@@ -18,11 +18,16 @@ from pyogrio import (
     vsi_rmtree,
     vsi_unlink,
 )
-from pyogrio._compat import GDAL_GE_38
+from pyogrio._compat import GDAL_GE_38, GDAL_GE_350
 from pyogrio._env import GDALEnv
 from pyogrio.errors import DataLayerError, DataSourceError
 from pyogrio.raw import read, write
-from pyogrio.tests.conftest import START_FID, prepare_testfile, requires_shapely
+from pyogrio.tests.conftest import (
+    DRIVERS,
+    START_FID,
+    prepare_testfile,
+    requires_shapely,
+)
 
 import pytest
 
@@ -456,6 +461,14 @@ def test_read_info(naturalearth_lowres):
 
 
 @pytest.mark.parametrize(
+    "naturalearth_lowres", [*DRIVERS.keys(), ".sqlite"], indirect=True
+)
+def test_read_info_encoding(naturalearth_lowres):
+    meta = read_info(naturalearth_lowres)
+    assert meta["encoding"].upper() == "UTF-8"
+
+
+@pytest.mark.parametrize(
     "testfile", ["naturalearth_lowres_vsimem", "naturalearth_lowres_vsi"]
 )
 def test_read_info_vsi(testfile, request):
@@ -570,10 +583,26 @@ def test_read_info_force_total_bounds(
         assert info["total_bounds"] is None
 
 
+def test_read_info_jsonfield(nested_geojson_file):
+    """Test if JSON fields types are returned correctly."""
+    meta = read_info(nested_geojson_file)
+    assert meta["ogr_types"] == ["OFTString", "OFTString"]
+    if GDAL_GE_350:
+        # OFSTJSON is only supported for GDAL >= 3.5
+        assert meta["ogr_subtypes"] == ["OFSTNone", "OFSTJSON"]
+    else:
+        assert meta["ogr_subtypes"] == ["OFSTNone", "OFSTNone"]
+
+
 def test_read_info_unspecified_layer_warning(data_dir):
     """Reading a multi-layer file without specifying a layer gives a warning."""
     with pytest.warns(UserWarning, match="More than one layer found "):
         read_info(data_dir / "sample.osm.pbf")
+
+
+def test_read_info_invalid_layer(naturalearth_lowres):
+    with pytest.raises(ValueError, match="'layer' parameter must be a str or int"):
+        read_bounds(naturalearth_lowres, layer=["list_arg_is_invalid"])
 
 
 def test_read_info_without_geometry(no_geometry_file):