+++ CHANGES.md
@@ -6,9 +6,26 @@
 
 -   pyogrio now requires a minimum version of Python >= 3.10 (#557).
 
+### Potentially breaking changes
+
+-   Return JSON fields (as identified by GDAL) as dicts/lists in `read_dataframe`;
+    these were previously returned as strings (#556).
+
+### Improvements
+
+-   Add listing of GDAL data types and subtypes to `read_info` (#556).
+-   Add support to read list fields without arrow (#558).
+
 ### Bug fixes
 
+-   Fix decode error reading an sqlite file on windows (#568).
 -   Fix wrong layername when creating .gpkg.zip file (#570).
+-   Fix segfault on providing an invalid value for `layer` in `read_info` (#564).
+
+### Packaging
+
+-   Add libkml driver to the wheels for more recent Linux platforms supported
+    by manylinux_2_28, MacOS, and Windows (#561).
 
 ## 0.11.1 (2025-08-02)